refactor(domain): centralize project path normalization
